python_agent/requirements.txt
requests
python-dotenv
openai
ipython 